+++ juriscraper/opinions/united_states/state/nycrimct.py
@@ -1,0 +1,5 @@
+from juriscraper.opinions.united_states.state import nytrial
+
+
+class Site(nytrial.Site):
+    court_regex = r"Cri?m(inal)? C[our]*t"