+++ juriscraper/opinions/united_states/territories/virginislands.py
@@ -52,4 +52,3 @@ class Site(OpinionSiteLinear):
                     "url": url,
                 }
             )
-            print(self.cases)